% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.R
\name{measurement}
\alias{measurement}
\title{measurement}
\usage{
measurement(a, b)
}
\description{
Find measurement
}
\examples{
# to determinate the measurement of a circuit without any gate
measurement(qc,1)
# to determinate the measurement of a circuit with a gate but without Cnot gate
measurement(qc,2)
# to determinate the measurement of a circuit with Cnot gate type (0,1)
measurement(qc,3)
# to determinate the measurement of a circuit with Cnot gate type (1,0)
measurement(qc,4)

}
