% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmeasurement.R
\name{plotMeasurement}
\alias{plotMeasurement}
\title{Plot measurement}
\usage{
plotMeasurement(a)
}
\description{
Plot the measurement
}
\examples{
# to plot the measurement of a circuit called qc

plotMeasurement(measurement(qc,1))



}
